requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==4.0.1
itypes==1.2.0
Jinja2==2.11.3
Markdown==3.3.4
MarkupSafe==1.1.1
mysqlclient==2.0.3
openapi-codec==1.3.2
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
uWSGI==2.0.19.1
zipp==3.4.1
alabaster==0.7.12
appdirs==1.4.4
asgiref==3.3.4
astor==0.8.1
attrs==20.3.0
autoflake==1.4
Babel==2.9.1
bandit==1.7.0
black==20.8b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
coreapi==2.3.3
coreschema==0.0.4
coverage==5.5
darglint==1.8.0
distlib==0.3.1
Django==3.2
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
doc8==0.8.1
docutils==0.17.1
eradicate==1.0
filelock==3.0.12
flake8==3.9.1
flake8-2020==1.6.0
flake8-bandit==2.1.2
flake8-broken-line==0.2.1
flake8-bugbear==19.8.0
flake8-builtins==1.5.3
flake8-commas==2.0.0
flake8-comprehensions==3.4.0
flake8-debugger==3.2.1
flake8-docstrings==1.6.0
flake8-eradicate==0.3.0
flake8-fixme==1.1.1
flake8-isort==3.0.1
flake8-logging-format==0.6.0
flake8-mutable==1.2.0
flake8-polyfill==1.0.2
flake8-quotes==2.1.2
flake8-rst-docstrings==0.0.12
flake8-string-format==0.2.3
flake8-variables-names==0.0.3
gitdb==4.0.7
GitPython==3.1.14
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
isort==4.3.21
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.24
mysqlclient==2.0.3
openapi-codec==1.3.2
packaging==20.9
pathspec==0.8.1
pbr==5.6.0
pep8-naming==0.9.1
pluggy==0.13.1
protobuf==3.15.8
py==1.10.0
pycodestyle==2.7.0
pydocstyle==5.1.1
pyflakes==2.3.1
Pygments==2.8.1
PyMySQL==1.0.2
pyparsing==2.4.7
pytest==6.2.3
pytest-cov==2.11.1
python-dev-tools==2020.9.10
pytz==2021.1
pyupgrade==2.13.0
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
restructuredtext-lint==1.3.2
simplejson==3.17.2
six==1.15.0
smmap==4.0.0
snowballstemmer==2.1.0
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
stevedore==3.3.0
testfixtures==6.17.1
tokenize-rt==4.1.0
toml==0.10.2
tox==3.23.0
tox-travis==0.12
typed-ast==1.4.3
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.4.4
wemake-python-styleguide==0.14.1
